reworked item attribute bonus detection to include tools info

+++ src/haven/ItemFilter.java
@@ -606,7 +606,7 @@ public class ItemFilter {
 	
 	@Override
 	public boolean matches(List<ItemInfo> info) {
-	    Map<Resource, Integer> bonuses = ItemInfo.getBonuses(info, null);
+	    Map<Resource, Double> bonuses = ItemInfo.getBonuses(info, null);
 	    if(text != null && text.length() >= 3) {
 		for (Resource res : bonuses.keySet()) {
 		    Resource.Tooltip tip = res.layer(Resource.tooltip);
